pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-tag"
version = "0.8.2"
description = "Automatically tag a branch based on commit message"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Matei-Marius Micu", email = "contact@mateimicu.com"}
]
requires-python = ">=3.9"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Development Status :: 3 - Alpha",
    "Environment :: MacOS X",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Natural Language :: English",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: System :: Software Distribution",
]
dependencies = [
    "gitpython>=3.1.18",
    "semantic_version>=2.8.5",
    "confuse>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "types-pyyaml",
]

[project.scripts]
auto-tag = "auto_tag.__main__:main"

[project.urls]
Homepage = "https://github.com/mateimicu/auto-tag"
Repository = "https://github.com/mateimicu/auto-tag"

[tool.setuptools.packages.find]
exclude = ["tests*"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["auto_tag"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "git.*",
    "semantic_version.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["auto_tag/tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"

[tool.coverage.run]
source = ["auto_tag"]
omit = ["auto_tag/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]
